requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
google-generativeai>=0.4.1
plotly>=5.18.0
altair>=5.2.0
matplotlib>=3.8.0